Tidy up the frontend app
DRY up the views using layouts and components.
